fix: Добавить статистику с подсказкой
