+++ frontend/src/components/Chart/Chart.module.css
@@ -1,0 +1,5 @@
+.Tooltip {
+    background-color: rgba(240, 240, 240, 0.8);
+    border: 2px solid #6a91ad;
+    padding: 0px 20px;
+}
